Update ping from 020927 to the latest s20121221.
Our current copy of ping is based on the 2002 snapshot plus a
few local modifications, mostly to get it to compile on Android
and to allow permissionless ping using the ping socket.

This patch preserves the local modifications (except where they
are obsolete) by applying the diff between 020927 and s2012221
to our local copy. As expected, diffing our copy with s2012221
shows our local modifications.

This patch also includes the ping6 source code, but no attempt
is made to compile it yet (more changes are needed for that).

Tested on crespo using ToT master. As expected, "adb shell ping"
works when the shell is not running as root and /system/bin/ping
is not suid.

(cherry-pick of b933d68e8e35f0a48ef826dda1afb0ce2890d6ab.)

Bug: 9469682
Change-Id: I0cad3c38c449b3ab2e791350458511125ffbca94